Fix navigation menu hover effect
This commit fixes an issue with the navigation menu's hover effect. Previously, the menu item wouldn't get larger and change color on hover due to a typo in the CSS selector. Specifically, the ':is(hover, :focus)' pseudo-class was replaced with ':is(:hover, :focus)' to correctly target the button element. As a result, users can now see a clear visual feedback when hovering over the navigation menu, improving the overall user experience.